refactor(data-model): fix issues after refactor
