+++ packages/data-access/src/index.ts
@@ -1,6 +1,4 @@
 
-export { default as Project } from './models/Project';
-export { default as MutationTestingReport } from './models/MutationTestingReport';
+export * from './models';
 export * from './mappers';
-export { Mapper } from './mappers/Mapper';
 export { MutationTestingReportService } from './services/MutationTestingReportService';
