+++ packages/website-frontend/package.json
@@ -19,32 +19,28 @@
     "url": "https://github.com/stryker-mutator/stryker-dashboard",
     "directory": "packages/website-frontend"
   },
-  "dependencies": {
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.800.0",
     "@angular/animations": "~8.0.0",
+    "@angular/cli": "~8.0.2",
     "@angular/common": "~8.0.0",
     "@angular/compiler": "~8.0.0",
+    "@angular/compiler-cli": "~8.0.0",
     "@angular/core": "~8.0.0",
     "@angular/forms": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
     "@angular/platform-browser": "~8.0.0",
     "@angular/platform-browser-dynamic": "~8.0.0",
     "@angular/router": "~8.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.1",
+    "@ng-bootstrap/ng-bootstrap": "~5.1.4",
     "@stryker-mutator/dashboard-common": "^0.3.1",
     "@stryker-mutator/dashboard-contract": "^0.3.1",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.4.0",
-    "zone.js": "~0.9.0"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "~8.0.2",
-    "@angular/compiler-cli": "~8.0.0",
-    "@angular/language-service": "~8.0.0",
     "@types/jasmine": "~3.3.12",
     "@types/jasminewd2": "~2.0.6",
     "@types/node": "~11.13.0",
+    "bootstrap": "^4.3.1",
     "codelyzer": "~5.0.0",
+    "core-js": "^2.5.4",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.1",
@@ -54,8 +50,10 @@
     "karma-jasmine-html-reporter": "^1.4.0",
     "mutation-testing-report-schema": "^1.2.0",
     "protractor": "~5.4.2",
+    "rxjs": "~6.4.0",
     "ts-node": "~8.0.3",
     "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "typescript": "~3.4.3",
+    "zone.js": "~0.9.0"
   }
 }
